Major changes for feeds being pulled.

+++ src/main/java/com/sqrfactor/model/Feed.java
@@ -67,7 +67,7 @@ public class Feed {
 	
 	public Feed(){}
 
-	public Feed(int feedId, int userId, int feedTypeId, String feedText, String feedPath, int feedActionId,
+	public Feed(long feedId, long userId, int feedTypeId, String feedText, String feedPath, int feedActionId,
 			long feedRefId, String placeName, String placeAddress, double placeLat, double placeLng, Date createdAt, Date modifiedAt) {
 		super();
 		this.feedId = feedId;
